@@ -11,31 +11,30 @@
 		"format": "prettier --write --plugin-search-dir=. ."
 	},
 	"devDependencies": {
-		"@rollup/plugin-dsv": "^3.0.4",
-		"@sveltejs/adapter-static": "^3.0.4",
-		"@sveltejs/kit": "^2.5.25",
-		"@sveltejs/vite-plugin-svelte": "3.1.2",
+		"@rollup/plugin-dsv": "^3.0.5",
+		"@sveltejs/adapter-static": "^3.0.5",
+		"@sveltejs/kit": "^2.7.1",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0-next.6",
 		"archieml": "^0.5.0",
 		"autoprefixer": "^10.4.20",
-		"bits-ui": "^0.21.13",
+		"bits-ui": "^0.21.16",
 		"d3": "^7.9.0",
 		"lodash.debounce": "^4.0.8",
-		"lucide-svelte": "^0.436.0",
-		"postcss": "^8.4.41",
+		"lucide-svelte": "^0.453.0",
+		"postcss": "^8.4.47",
 		"prettier": "^3.3.3",
-		"prettier-plugin-svelte": "^3.2.6",
-		"rollup": "^4.21.2",
+		"prettier-plugin-svelte": "^3.2.7",
+		"rollup": "^4.24.0",
 		"rollup-plugin-svg": "^2.0.0",
-		"style-dictionary": "^4.1.0",
+		"style-dictionary": "^4.1.3",
 		"svelte": "5.0.0-next.242",
-		"svelte-copy": "^1.4.2",
-		"svelte-preprocess": "^6.0.2",
-		"vite": "^5.3.4",
+		"svelte-preprocess": "^6.0.3",
+		"vite": "^5.4.9",
 		"vite-plugin-svgstring": "^1.0.0"
 	},
 	"type": "module",
 	"engines": {
-		"node": ">= 18.12.0"
+		"node": ">= 18.20.4"
 	},
 	"browserslist": "> 0.5%, last 4 versions, not ie <= 11, not ie_mob <= 11"
 }
