@@ -1,5 +1,5 @@
 <script>
-	export let number;
+	let { number } = $props();
 </script>
 
 <p>I am component A and my favorite number is {number}.</p>
