@@ -1,5 +1,5 @@
 <script>
-	export let name;
+	let { name } = $props();
 </script>
 
 <p>I am component B and my name is {name}.</p>
