@@ -11,28 +11,23 @@
 	 * optional params with defaults
 	 * <Scrolly root={null} top={0} bottom={0} increments={100}>
 	 */
-	import { onMount } from "svelte";
-	export let root = null;
-	export let top = 0;
-	export let bottom = 0;
-	export let increments = 100;
-	export let value = undefined;
+	
+	let {
+		root = null,
+		top = 0,
+		bottom = 0,
+		increments = 100,
+		value = $bindable(undefined),
+		children
+	} = $props();
 
-	const steps = [];
-	const threshold = [];
-
+	let steps = [];
+	let threshold = [];
 	let nodes = [];
 	let intersectionObservers = [];
-	let container;
+	let container = undefined;
 
-	$: top, bottom, update();
-
-	const update = () => {
-		if (!nodes.length) return;
-		nodes.forEach(createObserver);
-	};
-
-	const mostInView = () => {
+	function mostInView () {
 		let maxRatio = 0;
 		let maxIndex = 0;
 		for (let i = 0; i < steps.length; i++) {
@@ -46,7 +41,7 @@
 		else value = undefined;
 	};
 
-	const createObserver = (node, index) => {
+	function createObserver(node, index) {
 		const handleIntersect = (e) => {
 			const intersecting = e[0].isIntersecting;
 			const ratio = e[0].intersectionRatio;
@@ -64,17 +59,29 @@
 		const io = new IntersectionObserver(handleIntersect, options);
 		io.observe(node);
 		intersectionObservers[index] = io;
-	};
+	}
 
-	onMount(() => {
+	function update() {
+		if (!nodes.length) return;
+		nodes.forEach(createObserver);
+	}
+
+	$effect(() => {
 		for (let i = 0; i < increments + 1; i++) {
 			threshold.push(i / increments);
 		}
 		nodes = container.querySelectorAll(":scope > *:not(iframe)");
 		update();
 	});
+
+	$effect(() => {
+		top;
+		bottom;
+		update();
+	});
+
 </script>
 
 <div bind:this={container}>
-	<slot />
+	{@render children?.()}
 </div>
