@@ -1,5 +1,5 @@
-<script>
+<!-- <script>
 	import IgStory from "$components/demo/Demo.IgStory.svelte";
 </script>
 
-<IgStory />
+<IgStory /> -->
